+++ src/package.json
@@ -1,7 +1,7 @@
 {
   "name": "labortablo",
   "description": "Labortablo Web UI",
-  "version": "0.0.11",
+  "version": "0.0.12-0",
   "private": true,
   "scripts": {
     "start": "node ./index.js",
